requirements.txt
# Core dependencies
openai>=1.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
tqdm>=4.66.1
backoff>=2.2.1

# Database
motor>=3.3.2  # Async MongoDB driver
pymongo>=4.5.0
mongoengine>=0.27.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mongodb>=2.4.0
pytest-asyncio>=0.21.1

# Type checking
mypy>=1.5.1
types-PyYAML>=6.0.12.12

# Logging
rich>=13.5.2

# Data handling
pandas>=2.1.4
numpy>=1.26.3

# GitHub API
PyGithub>=2.1.1

# Development tools
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0

# HTTP and async
aiohttp==3.9.3 